+++ src/main/java/org/osiam/resources/scim/User.java
@@ -1189,6 +1189,10 @@ public final class User extends Resource implements Serializable {
             return this;
         }
 
+        /**
+         * @deprecated You should not need to set the ID with a client. Will be removed in 1.12 or 2.0.
+         */
+        @Deprecated
         @Override
         public Builder setId(String id) {
             super.setId(id);